+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Version 2.99.x and version 3.x will be available with a [pro license](http://www
 **Next release**
 
 * Added application specific quirks ([issue](/../../issues/1844))
+* Added intent for update check ([issue](/../../issues/1867))
 * Updated Simplified Chinese translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)